Horrible hack in PiInput.c and PiShortcurt.c to differentiate = from F4 and ; from F2. Whole kbdTable required rewriting Everything is working like the last Pi3 version. Anything beyond this point will be considered new features

+++ Src/Pi/PiShortcuts.h
@@ -48,7 +48,9 @@ void shortcutsSetDirectory(char* directory);
 
 Shortcuts* shortcutsCreate();
 void shortcutsDestroy(Shortcuts* shortcuts);
-void shortcutCheckDown(Shortcuts* s, int type, int mods, int keySym);
-void shortcutCheckUp(Shortcuts* s, int type, int mods, int keySym);
+//void shortcutCheckDown(Shortcuts* s, int type, int mods, int keySym);
+//void shortcutCheckUp(Shortcuts* s, int type, int mods, int keySym);
+void shortcutCheckDown(Shortcuts* s, int type, int mods, long keySym);
+void shortcutCheckUp(Shortcuts* s, int type, int mods, long keySym);
 
 #endif
